Add function to test FTP connection Fix $Evergreen command for x86 arch Remove ActiveSetup Add automatic download for transform file Improve error checking

+++ Adobe/Acrobat DC/Install.ps1
@@ -15,7 +15,8 @@ Write-Verbose -Message "Importing custom modules..." -Verbose
 [System.Net.WebRequest]::DefaultWebProxy.Credentials = [System.Net.CredentialCache]::DefaultCredentials
 
 # Install custom package providers list
-Foreach ($PackageProvider in $PackageProviders) {
+Foreach ($PackageProvider in $PackageProviders)
+{
     If (-not(Get-PackageProvider -ListAvailable -Name $PackageProvider -ErrorAction SilentlyContinue)) { Install-PackageProvider -Name $PackageProvider -Force }
 }
 
@@ -28,14 +29,17 @@ $PSGetVersion = [version](Find-PackageProvider -Name PowerShellGet).Version
 If ($PSGetVersion -gt $InstalledPSGetVersion) { Install-PackageProvider -Name PowerShellGet -Force }
 
 # Install and import custom modules list
-Foreach ($Module in $Modules) {
+Foreach ($Module in $Modules)
+{
     If (-not(Get-Module -ListAvailable -Name $Module)) { Install-Module -Name $Module -AllowClobber -Force | Import-Module -Name $Module -Force }
-    Else {
+    Else
+    {
         $InstalledModuleVersion = (Get-InstalledModule -Name $Module).Version
         $ModuleVersion = (Find-Module -Name $Module).Version
         $ModulePath = (Get-InstalledModule -Name $Module).InstalledLocation
         $ModulePath = (Get-Item -Path $ModulePath).Parent.FullName
-        If ([version]$ModuleVersion -gt [version]$InstalledModuleVersion) {
+        If ([version]$ModuleVersion -gt [version]$InstalledModuleVersion)
+        {
             Update-Module -Name $Module -Force
             Remove-Item -Path $ModulePath\$InstalledModuleVersion -Force -Recurse
         }
@@ -45,18 +49,22 @@ Foreach ($Module in $Modules) {
 Write-Verbose -Message "Custom modules were successfully imported!" -Verbose
 
 # Get the current script directory
-Function Get-ScriptDirectory {
+Function Get-ScriptDirectory
+{
     Remove-Variable appScriptDirectory
-    Try {
+    Try
+    {
         If ($psEditor) { Split-Path $psEditor.GetEditorContext().CurrentFile.Path } # Visual Studio Code Host
         ElseIf ($psISE) { Split-Path $psISE.CurrentFile.FullPath } # Windows PowerShell ISE Host
         ElseIf ($PSScriptRoot) { $PSScriptRoot } # Windows PowerShell 3.0-5.1
-        Else {
+        Else
+        {
             Write-Host -ForegroundColor Red "Cannot resolve script file's path"
             Exit 1
         }
     }
-    Catch {
+    Catch
+    {
         Write-Host -ForegroundColor Red "Caught Exception: $($Error[0].Exception.Message)"
         Exit 2
     }
@@ -71,6 +79,58 @@ $appScriptDirectory = Get-ScriptDirectory
 
 # Application related
 ##*===============================================
+Function Test-FTPConnection
+{
+    <#
+    .SYNOPSIS
+        The Test-FTPConnection function allows you to test an FTP connection.
+
+    .DESCRIPTION
+        The Test-FTPConnection function allows you to test an FTP connection.
+
+    .PARAMETER URL
+
+    .EXAMPLE
+        Test-FTPConnection -URL "ftp.adobe.com"
+
+        This will test the ftp connection for URL ftp.adobe.com.
+
+    .NOTES
+        Author: Jonathan Pitre
+        Twitter: @PitreJonathan
+    #>
+
+    [CmdletBinding()]
+    Param(
+        [Parameter(
+            Mandatory = $true,
+            ValueFromPipeline = $true)]
+        [String]$URL,
+
+        [Alias("RunAs")]
+        [System.Management.Automation.Credential()]
+        $Credential = [System.Management.Automation.PSCredential]::Empty
+
+    ) #Param
+
+    Process
+    {
+        If ($URL)
+        {
+            Write-Verbose -Message "One or more URL specified" -Verbose
+            Try
+            {
+                Test-NetConnection -ComputerName $URL -Port 21 | Select-Object -ExpandProperty TcpTestSucceeded
+            } #Try
+            Catch
+            {
+                Throw "Failed to connect to URL: $URL with error $_."
+                Break
+            } #Catch
+        } # If
+    } #Process
+}
+
 $appVendor = "Adobe"
 $appName = "Acrobat"
 $appShortVersion = "DC"
@@ -81,67 +141,108 @@ $appSetup = "AcroPro.msi"
 $appInstallParameters = "/QB"
 #$appParameters = "--tool=VolumeSerialize --generate --serial=1016-1899-8440-6413-0576-7429 --leid=V7{}AcrobatCont-12-Win-GM --regsuppress=ss --eulasuppress --stream" #--provfile Optional; path of the folder where prov.xml is created. If this parameter is not specified, prov.xml is created in the folder in which APTEE resides.
 $appAddParameters = "IGNOREVCRT64=1 EULA_ACCEPT=YES UPDATE_MODE=0 DISABLE_ARM_SERVICE_INSTALL=1 ROAMIDENTITY=1 ROAMLICENSING=1"
-$Evergreen = Get-EvergreenApp -Name AdobeAcrobat| Where-Object {$_.Product -eq "Acrobat" -and $_.Track -eq "DC"}
+$Evergreen = Get-EvergreenApp -Name AdobeAcrobat | Where-Object {$_.Product -eq "Acrobat" -and $_.Track -eq "DC" -and $_.Architecture -eq "x86"}
 $appVersion = $Evergreen.Version
 $appURLPatch = $Evergreen.URI
 $appPatch = Split-Path -Path $appURLPatch -Leaf
+$appTransformURL = "https://github.com/JonathanPitre/Apps/raw/master/Adobe/Acrobat DC/AcroPro.mst"
+$appTransform = Split-Path -Path $appTransformURL -Leaf
 $appURLADMX = "ftp://ftp.adobe.com/pub/adobe/acrobat/win/AcrobatDC/misc/AcrobatADMTemplate.zip"
 $appADMX = Split-Path -Path $appURLADMX -Leaf
 $appURLCustWiz = "ftp://ftp.adobe.com/pub/adobe/acrobat/win/AcrobatDC/misc/CustWiz2000920067_en_US_DC.exe"
 $appCustWiz = Split-Path -Path $appURLCustWiz -Leaf
 $appCustWizVersion = $appCustWiz.Trim("CustWiz").Trim("_en_US_DC.exe")
+$appFTP = Test-FTPConnection -URL "ftp.adobe.com"
 $appDestination = "${env:ProgramFiles(x86)}\$appVendor\$appName $appShortVersion\$appName"
 [boolean]$IsAppInstalled = [boolean](Get-InstalledApplication -Name "$appVendor $appName $appShortVersion")
 $appInstalledVersion = (Get-InstalledApplication -Name "$appVendor $appName $appShortVersion").DisplayVersion
 ##*===============================================
 
-If ([version]$appVersion -gt [version]$appInstalledVersion) {
+If ([version]$appVersion -gt [version]$appInstalledVersion)
+{
     Set-Location -Path $appScriptDirectory
 
     # Uninstall previous versions
     Get-Process -Name $appProcesses | Stop-Process -Force
-    If ($IsAppInstalled) {
+    If (($IsAppInstalled) -and (Test-Path -Path $appSetup))
+    {
         Write-Log -Message "Uninstalling previous versions..." -Severity 1 -LogType CMTrace -WriteHost $True
         Remove-MSIApplications -Name "$appVendor $appName $appShortVersion"
     }
 
+    # Download latest policy definitions
+    If ($appFTP)
+    {
+        Write-Log -Message "Downloading $appVendor $appName $appShortVersion ADMX templates..." -Severity 1 -LogType CMTrace -WriteHost $True
+        Invoke-WebRequest -UseBasicParsing -Uri $appURLADMX -OutFile $appADMX
+        New-Folder -Path "$appScriptDirectory\PolicyDefinitions"
+        Expand-Archive -Path $appADMX -DestinationPath "$appScriptDirectory\PolicyDefinitions" -Force
+        Remove-File -Path $appADMX, $appScriptDirectory\PolicyDefinitions\$appName$($appShortVersion).adm
+    }
+    Else
+    {
+        Write-Log -Message "FTP URL is unreachable, latest policy definitions files won't be downloaded." -Severity 2 -LogType CMTrace -WriteHost $True
+    }
+
     # Download latest Adobe Acrobat Customization Wizard DC
-    If (-Not(Test-Path -Path $appScriptDirectory\$appCustWiz)) {
+    If (-Not(Test-Path -Path $appScriptDirectory\$appCustWiz) -and ($appFTP))
+    {
         Write-Log -Message "Downloading $appVendor $appName Custimization Wizard $appShortVersion $appCustWizVersion..." -Severity 1 -LogType CMTrace -WriteHost $True
         Invoke-WebRequest -UseBasicParsing -Uri $appURLCustWiz -OutFile $appCustWiz
     }
-    Else {
+    ElseIf (-not($appFTP))
+    {
+        Write-Log -Message "FTP URL is unreachable, $appVendor $appName Custimization Wizard $appShortVersion won't be downloaded. " -Severity 2 -LogType CMTrace -WriteHost $True
+    }
+    Else
+    {
         Write-Log -Message "File(s) already exists, download was skipped." -Severity 1 -LogType CMTrace -WriteHost $True
     }
 
-    # Download latest policy definitions
-    Write-Log -Message "Downloading $appVendor $appName $appShortVersion ADMX templates..." -Severity 1 -LogType CMTrace -WriteHost $True
-    Invoke-WebRequest -UseBasicParsing -Uri $appURLADMX -OutFile $appADMX
-    New-Folder -Path "$appScriptDirectory\PolicyDefinitions"
-    Expand-Archive -Path $appADMX -DestinationPath "$appScriptDirectory\PolicyDefinitions" -Force
-    Remove-File -Path $appADMX, $appScriptDirectory\PolicyDefinitions\$appName$($appShortVersion).adm
-
     # Download latest patch file
-    If (-Not(Test-Path -Path $appScriptDirectory\$appPatch)) {
+    If (-Not(Test-Path -Path $appScriptDirectory\$appPatch))
+    {
         Write-Log -Message "Downloading $appVendor $appName $appShortVersion $appVersion patch..." -Severity 1 -LogType CMTrace -WriteHost $True
         Invoke-WebRequest -UseBasicParsing -Uri $appURLPatch -OutFile $appPatch
-        If (Test-Path -Path $appScriptDirectory\$appPatch) {
-            Set-IniValue -FilePath $appScriptDirectory\setup.ini -Section "Startup" -Key "CmdLine" -Value "/sPB /rs"
+        If ((Test-Path -Path $appScriptDirectory\$appPatch) -and (Test-Path -Path $appScriptDirectory\$appPatch\setup.ini))
+        {
+            Set-IniValue -FilePath $appScriptDirectory\setup.ini -Section "Startup" -Key "CmdLine" -Value "/sPB /rs /msi $appAddParameters"
             Set-IniValue -FilePath $appScriptDirectory\setup.ini -Section "Product" -Key "CmdLine" -Value "TRANSFORMS=`"$appTransform`""
             Set-IniValue -FilePath $appScriptDirectory\setup.ini -Section "Product" -Key "PATCH" -Value $appPatch
         }
     }
-    Else {
+    Else
+    {
         Write-Log -Message "File(s) already exists, download was skipped." -Severity 1 -LogType CMTrace -WriteHost $True
     }
 
-    # Install latest version
-    If ((Test-Path -Path $appScriptDirectory\$appSetup) -and (Test-Path -Path $appScriptDirectory\$appTransform)) {
+
+    If ((Test-Path -Path $appScriptDirectory\$appSetup) -and (Test-Path -Path $appScriptDirectory\$appPatch))
+    {
+        # Download required transform file
+        If (-Not(Test-Path -Path $appScriptDirectory\$appTransform))
+        {
+            Write-Log -Message "Downloading $appVendor $appName $appShortVersion Transform.." -Severity 1 -LogType CMTrace -WriteHost $True
+            Invoke-WebRequest -UseBasicParsing -Uri $appTransformURL -OutFile $appScriptDirectory\$appTransform
+        }
+        Else
+        {
+            Write-Log -Message "File(s) already exists, download was skipped." -Severity 1 -LogType CMTrace -WriteHost $True
+        }
+        # Install latest version
         Write-Log -Message "Installing $appVendor $appName $appShortVersion $appVersion..." -Severity 1 -LogType CMTrace -WriteHost $True
         Execute-MSI -Action Install -Path $appSetup -Transform $appTransform -Parameters $appInstallParameters -AddParameters $appAddParameters -Patch $appPatch -SkipMSIAlreadyInstalledCheck
     }
-    Else {
-        Write-Log -Message "Setup file(s) are missing." -Severity 1 -LogType CMTrace -WriteHost $True
+    ElseIf (($IsAppInstalled) -and (Test-Path -Path $appScriptDirectory\$appPatch))
+    {
+        # Install latest patch
+        Write-Log -Message "Setup file(s) are missing, MSP file will be installed." -Severity 1 -LogType CMTrace -WriteHost $True
+        Write-Log -Message "Installing $appVendor $appName $appShortVersion $appVersion..." -Severity 1 -LogType CMTrace -WriteHost $True
+        Execute-MSP -Path $appPatch -Parameters $appInstallParameters
+    }
+    Else
+    {
+        Write-Log -Message "Setup file(s) are missing" -Severity 2 -LogType CMTrace -WriteHost $True
         Exit-Script
     }
 
@@ -166,7 +267,11 @@ If ([version]$appVersion -gt [version]$appInstalledVersion) {
     Get-Process -Name $appProcesses | Stop-Process -Force
     #>
 
-    If (-Not(Get-InstalledApplication -Name "Adobe Creative Cloud")) {
+    # Remove Active Setup
+    Remove-RegistryKey -Key "HKLM:\SOFTWARE\WOW6432Node\Microsoft\Active Setup\Installed Components\{AC76BA86-0000-0000-7760-7E8A45000000}" -Name "StubPath"
+
+    If (-Not(Get-InstalledApplication -Name "Adobe Creative Cloud"))
+    {
         Write-Log -Message "Adobe Creative Cloud must be installed in order for $appVendor $appName $appShortVersion licensing to work!" -Severity 2 -LogType CMTrace -WriteHost $True
     }
 
@@ -175,6 +280,7 @@ If ([version]$appVersion -gt [version]$appInstalledVersion) {
 
     Write-Log -Message "$appVendor $appName $appShortVersion $appVersion was installed successfully!" -Severity 1 -LogType CMTrace -WriteHost $True
 }
-Else {
+Else
+{
     Write-Log -Message "$appVendor $appName $appShortVersion $appInstalledVersion is already installed." -Severity 1 -LogType CMTrace -WriteHost $True
 }
